added a way to update when a course is added the NEW tome way
