+++ src/components/semesterList.tsx
@@ -1,27 +1,24 @@
 import React from "react";
 import { Semester } from "../interfaces/semester";
-import { SemesterView } from "./SemesterView";
+import { SemesterView } from "./semesterView";
 import "../styleSheets/multipleSemesterTable.css";
 import { Course } from "../interfaces/course";
+import { Plan } from "../interfaces/plan";
 
 export function SemesterList({
     semesters,
     deleteSemester,
     editSemester,
-    updateCoursePlan,
     planId,
     updateEditedSemester,
     updateEditedCourse,
-    updateDeletedCourse
+    updateDeletedCourse,
+    plans,
+    setPlans
 }: {
     semesters: Semester[];
     deleteSemester: (id: string) => void;
     editSemester: (id: string, newSemester: Semester) => void;
-    updateCoursePlan: (
-        planId: string,
-        semId: string,
-        newCourse: Course
-    ) => void;
     planId: string;
     updateEditedSemester: (
         planId: string,
@@ -42,6 +39,8 @@ export function SemesterList({
         semId: string,
         courseCode: string
     ) => void;
+    plans: Plan[];
+    setPlans: (p: Plan[]) => void;
 }): JSX.Element {
     //make sure the tables stay in a nice format/gridlike
     //calls semesterview which will determine if we're in editing mode
@@ -49,11 +48,12 @@ export function SemesterList({
         <div className="container">
             {semesters.map((sem: Semester) => (
                 <SemesterView
+                    plans={plans}
+                    setPlans={setPlans}
                     key={sem.id}
                     semester={sem}
                     editSemester={editSemester}
                     deleteSemester={deleteSemester}
-                    updateCoursePlan={updateCoursePlan}
                     planId={planId}
                     updateEditedCourse={updateEditedCourse}
                     updateEditedSemester={updateEditedSemester}
